"Fix" base64 encoding problems of legacy library.
In the previous commits the library was altered to use the window.atob
and window.btoa functions. But the bit streams and base64 encoded
strings that are created by the legacy PGP library are not compatible
with the polyfill. The polyfill is needed in banner.api.js so we have to
bake in two libraries.