Allow for empty period and value selection
Period and value radio buttons can now be empty and are checked
client-side (the server-side check of period assumes "once" when no
period is checked and does not mark it as invalid).

The checks abort the sending of data to the server.
The error message for the Error objects is ignored, for i18N and alerts
the `alertMessage` property is used instead. This is because
validateAndSetPeriod has already an alert.